Create primary and auxiliary folders
