+++ src/components/primary/Header.tsx
@@ -1,41 +1,41 @@
 import { MoonIcon, SunIcon } from "@heroicons/react/20/solid";
-import { useTheme } from "../state/theme";
+import { useTheme } from "../../state/theme";
 import { signOut, useSession } from "next-auth/react";
 import Link from "next/link";
 import { useRouter } from "next/router";
 
 export function Header() {
-  const { data: session, status } = useSession();
+  // const { data: session, status } = useSession();
   const router = useRouter();
   const isActive: (pathname: string) => boolean = (pathname) =>
     router.pathname === pathname;
 
   let right = null;
-  if (session) {
-    right = (
-      <div className="flex items-center gap-x-2 text-xs">
-        <div className="rounded-full py-2 px-3 ring-1 ring-gray-950/5">
-          {session.user?.name ?? "No name"}
-        </div>
-        <button
-          onClick={() => signOut()}
-          className="px-2 py1 text-center hover:text-gray-500"
-        >
-          Sign out
-        </button>
-      </div>
-    );
-  } else if (status === "loading") {
-    right = <div className="text-xs flex items-center">Loading...</div>;
-  } else {
-    right = (
-      <div className="text-xs flex items-center">
-        <Link href="/api/auth/signin" data-active={isActive("/signup")}>
-          Log in
-        </Link>
-      </div>
-    );
-  }
+  // if (session) {
+  //   right = (
+  //     <div className="flex items-center gap-x-2 text-xs">
+  //       <div className="rounded-full py-2 px-3 ring-1 ring-gray-950/5">
+  //         {session.user?.name ?? "No name"}
+  //       </div>
+  //       <button
+  //         onClick={() => signOut()}
+  //         className="px-2 py1 text-center hover:text-gray-500"
+  //       >
+  //         Sign out
+  //       </button>
+  //     </div>
+  //   );
+  // } else if (status === "loading") {
+  //   right = <div className="text-xs flex items-center">Loading...</div>;
+  // } else {
+  //   right = (
+  //     <div className="text-xs flex items-center">
+  //       <Link href="/api/auth/signin" data-active={isActive("/signup")}>
+  //         Log in
+  //       </Link>
+  //     </div>
+  //   );
+  // }
 
   return (
     <nav className="h-12 px-8 lg:px-24 py-2 fixed top-0 w-full bg-white dark:bg-gray-950 z-10 border-b border-gray-200">
