+++ src/pages/index.tsx
@@ -1,9 +1,9 @@
 import { prisma } from "@/lib/prisma";
 import { Verse } from "@prisma/client";
 import { Inter } from "next/font/google";
-import { Header } from "../components/Header";
-import { Verses } from "../components/Verses";
-import { VerseMemoryTestModal } from "../components/VerseMemoryTestModal";
+import { Header } from "../components/primary/Header";
+import { Verses } from "../components/primary/Verses";
+import { AuxVerseMemoryTestModal } from "../components/auxiliary/AuxVerseMemoryTestModal";
 import { useDaoStore } from "../state/store";
 
 const inter = Inter({ subsets: ["latin"] });
@@ -14,7 +14,7 @@ export default function Home({ verses }: { verses: Verse[] }) {
       <Header />
       <main className="px-8 pb-16 lg:px-24 lg:pb-24 pt-4 lg:pt-8 mt-12">
         <div className="z-10 w-full m-auto max-w-xl items-center justify-between font-mono text-sm">
-          <VerseMemoryTestModal />
+          <AuxVerseMemoryTestModal />
           <Verses verses={verses} />
         </div>
       </main>
